pyproject.toml
[project]
name = "metrologet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]

[tool.uv.workspace]
members = [
    "protocol_generator",
]
